modify all of files and moved them to correct position
