+++ deployments/docker-compose-debug.yml
@@ -5,16 +5,17 @@ version: "3.5"
 services:
   server:
     build:
-      dockerfile: ./Dockerfile.debug
-      context: .
-    image: remotedebug:latest
+      context: ../
+      dockerfile: build/debuger/Dockerfile.debug
+    image: mirchen/remotedebug-debug:latest
     working_dir: /appdebug
     volumes:
-      - .:/appdebug
+      - ../:/appdebug
     ports:
       # Golang Server export 5000 Port
       - 5000:5000
       # According delve "Debug Port" default is 6379, make Docker port forwarding from 6379 to 2345 for vscode debug mode listen
       - 2345:6379
     # Run container entry point is dlv debug mode, listen dvl default port
-    entrypoint: dlv debug --headless --listen=:6379 --log --api-version=2 -- daemon
+    # entrypoint: sleep 60
+    entrypoint: dlv debug ./cmd/basicapp/ --headless --listen=:6379 --log --api-version=2 -- daemon
