+++ deployments/docker-compose.yml
@@ -1,0 +1,17 @@
+# docker-compose.yaml
+
+version: "3.5"
+
+services:
+  server:
+    build:
+      context: ../
+      dockerfile: build/builder/Dockerfile
+    image: mirchen/remotedebug:latest
+    working_dir: /go/src/app
+    ports:
+      # Golang Server export 5000 Port
+      - 5000:5000
+    # Run container entry point is dlv debug mode, listen dvl default port
+    entrypoint: app-server daemon
+
