+++ scripts/startdebug.sh
@@ -1,0 +1,4 @@
+until [ -f ./__debug_bin ] 
+do 
+    sleep 1 
+done
